requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
bcrypt==5.0.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
filelock==3.20.0
firebase_admin==7.1.0
fonttools==4.60.1
fsspec==2025.9.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
panns-inference==0.1.1
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
torchaudio==2.9.0
torchlibrosa==0.1.0
torchvision==0.24.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
